+++ packages/microservices/client/client-mqtt.ts
@@ -1,9 +1,9 @@
 import { Logger } from '@nestjs/common/services/logger.service';
 import { loadPackage } from '@nestjs/common/utils/load-package.util';
+import { ERROR_EVENT, MESSAGE_EVENT, MQTT_DEFAULT_URL } from '../constants';
 import { MqttClient } from '../external/mqtt-client.interface';
+import { MqttOptions, PacketId, ReadPacket, WritePacket } from '../interfaces';
 import { ClientOptions } from '../interfaces/client-metadata.interface';
-import { ERROR_EVENT, MESSAGE_EVENT, MQTT_DEFAULT_URL } from './../constants';
-import { MqttOptions, PacketId, ReadPacket, WritePacket } from './../interfaces';
 import { ClientProxy } from './client-proxy';
 import { ECONNREFUSED } from './constants';
