+++ packages/microservices/client/client-proxy.ts
@@ -2,8 +2,8 @@ import { isNil } from '@nestjs/common/utils/shared.utils';
 import { defer, fromEvent, merge, Observable, Observer, throwError as _throw } from 'rxjs';
 import { map, mergeMap, take } from 'rxjs/operators';
 import { CONNECT_EVENT, ERROR_EVENT } from '../constants';
-import { InvalidMessageException } from '../exceptions/invalid-message.exception';
-import { ClientOptions, PacketId, ReadPacket, WritePacket } from './../interfaces';
+import { InvalidMessageException } from '../exceptions/errors/invalid-message.exception';
+import { ClientOptions, PacketId, ReadPacket, WritePacket } from '../interfaces';
 
 export abstract class ClientProxy {
   public abstract connect(): Promise<any>;
