+++ packages/microservices/client/client-tcp.ts
@@ -2,9 +2,9 @@ import { Logger } from '@nestjs/common';
 import * as JsonSocket from 'json-socket';
 import * as net from 'net';
 import { tap } from 'rxjs/operators';
+import { CLOSE_EVENT, ERROR_EVENT, MESSAGE_EVENT, TCP_DEFAULT_HOST, TCP_DEFAULT_PORT } from '../constants';
+import { PacketId, ReadPacket, WritePacket } from '../interfaces';
 import { ClientOptions, TcpClientOptions } from '../interfaces/client-metadata.interface';
-import { CLOSE_EVENT, ERROR_EVENT, MESSAGE_EVENT, TCP_DEFAULT_HOST, TCP_DEFAULT_PORT } from './../constants';
-import { PacketId, ReadPacket, WritePacket } from './../interfaces';
 import { ClientProxy } from './client-proxy';
 import { ECONNREFUSED } from './constants';
