+++ packages/microservices/exceptions/errors/invalid-grpc-service.exception.ts
