+++ packages/microservices/exceptions/errors/invalid-message.exception.ts
