+++ packages/microservices/exceptions/errors/invalid-proto-definition.exception.ts
